Change hw.physmem and hw.usermem to unsigned long like they used to be in the original hardwired sysctl implementation.
The buf size calculator still overflows an integer on machines with large
KVA (eg: ia64) where the number of pages does not fit into an int.  Use
'long' there.

Change Maxmem and physmem and related variables to 'long', mostly for
completeness.  Machines are not likely to overflow 'int' pages in the
near term, but then again, 640K ought to be enough for anybody.  This
comes for free on 32 bit machines, so why not?

+++ sys/alpha/include/md_var.h
@@ -36,7 +36,7 @@
 extern	char	sigcode[];
 extern	char	esigcode[];
 extern	int	szsigcode;
-extern	int	Maxmem;
+extern	long	Maxmem;
 extern	int	busdma_swi_pending;
 
 struct fpreg;
